Day 5 Complete

+++ BusinessServices/ProductServices.cs
@@ -1,70 +1,66 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using System.Transactions;
 using AutoMapper;
 using BusinessEntities;
 using DataModel;
-using DataModel.UnitOfWork; 
+using DataModel.UnitOfWork;
 
 namespace BusinessServices
 {
-     /// <summary>     
-     /// Offers services for product specific CRUD operations     
-     /// </summary>    
-     public class ProductServices: IProductServices
+    /// <summary>
+    /// Offers services for product specific CRUD operations
+    /// </summary>
+    public class ProductServices:IProductServices
     {
-        private readonly UnitOfWork _unitOfWork; 
- 
-        /// <summary>         
-        /// Public constructor.         
-        /// </summary>         
-        public ProductServices(UnitOfWork unitOfWork) 
+        private readonly UnitOfWork _unitOfWork;
+
+        /// <summary>
+        /// Public constructor.
+        /// </summary>
+        public ProductServices(UnitOfWork unitOfWork)
         {
             _unitOfWork = unitOfWork;
         }
 
-        /// <summary>         
-        /// Fetches product details by id         
-        /// </summary>         
-        /// <param name="productId"></param>         
-        /// <returns></returns>         
+        /// <summary>
+        /// Fetches product details by id
+        /// </summary>
+        /// <param name="productId"></param>
+        /// <returns></returns>
         public BusinessEntities.ProductEntity GetProductById(int productId)
         {
             var product = _unitOfWork.ProductRepository.GetByID(productId);
             if (product != null)
             {
-                //Mapper.CreateMap<Product, ProductEntity>();
+                Mapper.CreateMap<Product, ProductEntity>();
                 var productModel = Mapper.Map<Product, ProductEntity>(product);
                 return productModel;
             }
             return null;
-        } 
+        }
 
-        /// <summary>         
-        /// Fetches all the products.         
-        /// </summary>         
-        /// <returns></returns>         
+        /// <summary>
+        /// Fetches all the products.
+        /// </summary>
+        /// <returns></returns>
         public IEnumerable<BusinessEntities.ProductEntity> GetAllProducts()
         {
             var products = _unitOfWork.ProductRepository.GetAll().ToList();
             if (products.Any())
             {
-                //Mapper.CreateMap<Product, ProductEntity>();
-                System.Diagnostics.Debug.WriteLine("There are products in here but I messed up automapper");
+                Mapper.CreateMap<Product, ProductEntity>();
                 var productsModel = Mapper.Map<List<Product>, List<ProductEntity>>(products);
                 return productsModel;
             }
             return null;
-        } 
+        }
 
-        /// <summary>         
-        /// Creates a product         
-        /// </summary>         
-        /// <param name="productEntity"></param>         
-        /// <returns></returns>         
+        /// <summary>
+        /// Creates a product
+        /// </summary>
+        /// <param name="productEntity"></param>
+        /// <returns></returns>
         public int CreateProduct(BusinessEntities.ProductEntity productEntity)
         {
             using (var scope = new TransactionScope())
@@ -73,23 +69,19 @@ namespace BusinessServices
                 {
                     ProductName = productEntity.ProductName
                 };
-
                 _unitOfWork.ProductRepository.Insert(product);
                 _unitOfWork.Save();
                 scope.Complete();
                 return product.ProductId;
             }
-        } 
+        }
 
-
-
-
-        /// <summary>         
-        /// Updates a product         
-        /// </summary>         
-        /// <param name="productId"></param>         
-        /// <param name="productEntity"></param>         
-        /// <returns></returns>         
+        /// <summary>
+        /// Updates a product
+        /// </summary>
+        /// <param name="productId"></param>
+        /// <param name="productEntity"></param>
+        /// <returns></returns>
         public bool UpdateProduct(int productId, BusinessEntities.ProductEntity productEntity)
         {
             var success = false;
@@ -109,13 +101,13 @@ namespace BusinessServices
                 }
             }
             return success;
-        } 
+        }
 
-        /// <summary>         
-        /// Deletes a particular product         
-        /// </summary>         
-        /// <param name="productId"></param>         
-        /// <returns></returns>         
+        /// <summary>
+        /// Deletes a particular product
+        /// </summary>
+        /// <param name="productId"></param>
+        /// <returns></returns>
         public bool DeleteProduct(int productId)
         {
             var success = false;
@@ -123,8 +115,10 @@ namespace BusinessServices
             {
                 using (var scope = new TransactionScope())
                 {
-                    var product = _unitOfWork.ProductRepository.GetByID(productId); if (product != null)
+                    var product = _unitOfWork.ProductRepository.GetByID(productId);
+                    if (product != null)
                     {
+
                         _unitOfWork.ProductRepository.Delete(product);
                         _unitOfWork.Save();
                         scope.Complete();
@@ -135,5 +129,4 @@ namespace BusinessServices
             return success;
         }
     }
-} 
-
+}
